v0.6.0

+++ src/LuminoCore/src/Base/CRTHelper.h
@@ -6,6 +6,7 @@
 
 #if defined(LN_GNUC) || defined(__CYGWIN__)
 
+#include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
@@ -84,7 +85,7 @@ inline errno_t mbstowcs_s( size_t *pConvertedChars, wchar_t *wcstr, size_t sizeI
 	if (mbstr == NULL)
 		return EINVAL;
 		
-	int len = strlen(mbstr);
+	size_t len = strlen(mbstr);
 	if (len > count)
 		return ERANGE;
 	
@@ -199,7 +200,7 @@ inline errno_t strcpy_s(
 	if (!strDestination || !strSource) {
 		return EINVAL;
 	}
-	int len = strlen(strSource);
+	size_t len = strlen(strSource);
 	if (numberOfElements < len + 1) {
 		return ERANGE;
 	}
